+++ encode_fp_int_strided.cu
@@ -54,7 +54,7 @@ struct uint_transpose {
 };
 
 template <class real_t, class real_const_ptr, int32_t GRID_X, int32_t GRID_Y, int32_t BLOCK_THREADS, int32_t COMPLEX, int32_t BASE, int32_t ORDER>
-__global__ void encode_fp_strided_i8(int32_t order, int32_t M, int32_t N, real_const_ptr A, int32_t lda, const int32_t* __restrict__ vec_expon, uint32_t* __restrict__ inA, int32_t ldi, int32_t strideI) {
+__global__ void encode_fp_strided_i8(int32_t order, int32_t M, int32_t N, real_const_ptr A, int32_t lda, const int32_t* __restrict__ vec_expon, uint32_t* __restrict__ inA, int32_t ldi, uint64_t strideI) {
   constexpr int32_t items = 1 << COMPLEX;
   constexpr int32_t elements_block = items * BLOCK_THREADS;
   constexpr int32_t elements = GRID_Y * elements_block;
@@ -72,7 +72,7 @@ __global__ void encode_fp_strided_i8(int32_t order, int32_t M, int32_t N, real_c
   uint_transpose trans_f;
 
   for (int32_t col = blockIdx.y; col < N; col += GRID_X) {
-    real_const_ptr A_i = &A[col * lda];
+    real_const_ptr A_i = &A[uint64_t(col) * uint64_t(lda)];
     int32_t vec_e = vec_expon[col];
 
     for (int32_t i = block_offset; i < M1; i += elements) {
@@ -94,7 +94,7 @@ __global__ void encode_fp_strided_i8(int32_t order, int32_t M, int32_t N, real_c
       
       uint32_t* inA_i = &inA[(i >> COMPLEX) + col * ldi];
       for (int32_t k = 0; k < order; ++k)
-        block_store.Store(&inA_i[k * strideI], *(uint32_t(*)[1])(&threadB[k]));
+        block_store.Store(&inA_i[uint64_t(k) * strideI], *(uint32_t(*)[1])(&threadB[k]));
     }
 
     if (0 < M2 && blockIdx.x == 0) {
@@ -114,10 +114,10 @@ __global__ void encode_fp_strided_i8(int32_t order, int32_t M, int32_t N, real_c
       for (int32_t k = 0; k < ORDER; ++k)
         trans_f(*(uint32_t(*)[items])(&threadB[k * items]));
       
-      uint32_t* inA_i = &inA[(M1 >> COMPLEX) + col * ldi];
+      uint32_t* inA_i = &inA[uint64_t(M1 >> COMPLEX) + uint64_t(col) * uint64_t(ldi)];
       int32_t len_M2 = (M2 + items - 1) >> COMPLEX;
       for (int32_t k = 0; k < order; ++k)
-        block_store.Store(&inA_i[k * strideI], *(uint32_t(*)[1])(&threadB[k]), len_M2);
+        block_store.Store(&inA_i[uint64_t(k) * strideI], *(uint32_t(*)[1])(&threadB[k]), len_M2);
     }
   }
 }
@@ -126,22 +126,22 @@ constexpr int32_t block_threads = 128;
 constexpr int32_t grid_x = 32;
 constexpr int32_t grid_y = 64;
 
-void internal::int8::encode_f64(cudaStream_t stream, int32_t order, int32_t M, int32_t N, const double* A, int32_t lda, const int32_t* vec_expon, int8_t* inA, int32_t ldi, int32_t strideI) {
+void internal::int8::encode_f64(cudaStream_t stream, int32_t order, int32_t M, int32_t N, const double* A, int32_t lda, const int32_t* vec_expon, int8_t* inA, int32_t ldi, uint64_t strideI) {
   encode_fp_strided_i8 <double, const double* __restrict__, grid_x, grid_y, block_threads, 2, exp_base, order_max>
     <<< dim3(grid_y, grid_x, 1), block_threads, 0, stream >>> (order, M, N, (const double*)A, lda, vec_expon, (uint32_t*)inA, ldi >> 2, strideI >> 2);
 }
 
-void internal::int8::encode_f32(cudaStream_t stream, int32_t order, int32_t M, int32_t N, const float* A, int32_t lda, const int32_t* vec_expon, int8_t* inA, int32_t ldi, int32_t strideI) {
+void internal::int8::encode_f32(cudaStream_t stream, int32_t order, int32_t M, int32_t N, const float* A, int32_t lda, const int32_t* vec_expon, int8_t* inA, int32_t ldi, uint64_t strideI) {
   encode_fp_strided_i8 <float, const float* __restrict__, grid_x, grid_y, block_threads, 2, exp_base, order_max>
     <<< dim3(grid_y, grid_x, 1), block_threads, 0, stream >>> (order, M, N, (const float*)A, lda, vec_expon, (uint32_t*)inA, ldi >> 2, strideI >> 2);
 }
 
-void internal::int8::encode_cf64(cudaStream_t stream, int32_t order, int32_t M, int32_t N, const std::complex<double>* A, int32_t lda, const int32_t* vec_expon, int8_t* inA, int32_t ldi, int32_t strideI) {
+void internal::int8::encode_cf64(cudaStream_t stream, int32_t order, int32_t M, int32_t N, const std::complex<double>* A, int32_t lda, const int32_t* vec_expon, int8_t* inA, int32_t ldi, uint64_t strideI) {
   encode_fp_strided_i8 <double, const double* __restrict__, grid_x, grid_y, block_threads, 3, exp_base, order_max>
     <<< dim3(grid_y, grid_x, 1), block_threads, 0, stream >>> (order * 2, M * 2, N, (const double*)A, lda * 2, vec_expon, (uint32_t*)inA, ldi >> 2, strideI >> 2);
 }
 
-void internal::int8::encode_cf32(cudaStream_t stream, int32_t order, int32_t M, int32_t N, const std::complex<float>* A, int32_t lda, const int32_t* vec_expon, int8_t* inA, int32_t ldi, int32_t strideI) {
+void internal::int8::encode_cf32(cudaStream_t stream, int32_t order, int32_t M, int32_t N, const std::complex<float>* A, int32_t lda, const int32_t* vec_expon, int8_t* inA, int32_t ldi, uint64_t strideI) {
   encode_fp_strided_i8 <float, const float* __restrict__, grid_x, grid_y, block_threads, 3, exp_base, order_max>
     <<< dim3(grid_y, grid_x, 1), block_threads, 0, stream >>> (order * 2, M * 2, N, (const float*)A, lda * 2, vec_expon, (uint32_t*)inA, ldi >> 2, strideI >> 2);
 }
